Выносит слайдер в отдельный модуль
