+++ js/main.js
@@ -1,4 +1,5 @@
 import { renderUsersPictures } from './thumbnails.js';
 import './uploadForm.js';
+import './slider.js';
 
 renderUsersPictures();
